pyproject.toml
[tool.poetry]
name = "sandbox-flask"
version = "0.1.0"
description = "flask test app"
authors = ["Suzuki Masaya <ikachang-kawaii@nifty.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^1.1.2"
uWSGI = "^2.0.19"
Flask-SQLAlchemy = "^2.5.1"
pydantic = "^1.8.1"
email-validator = "^1.1.2"
psycopg2 = "^2.8.6"
injector = "^0.18.4"
Flask-Injector = "^0.12.3"

[tool.poetry.dev-dependencies]
mypy = "^0.812"
pytest = "^6.2.2"
flake8 = "^3.9.0"
black = "^20.8b1"
isort = "^5.8.0"
pysen = "^0.9.1"
pre-commit = "^2.12.0"
pytest-mock = "^3.5.1"
pytest-cov = "^2.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pysen]
version = "0.9"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = true
mypy_preset = "strict"
line_length = 88
py_version = "py38"

[[tool.pysen.lint.mypy_targets]]
  paths = ["app"]

